+++ server/src/routes/index.js
@@ -2,15 +2,16 @@
 
 const router = require('express').Router();
 const auth = require('../middlewares/auth');
+const franja = require('../middlewares/franjaHoraria');
 
 router.use('/auth', require('./auth.routes'));
-router.use('/usuarios', auth, require('./usuario.routes'));
-router.use('/accesos', auth, require('./accesos.routes'));
-router.use('/roles', auth, require('./rol.routes'));
-router.use('/catalogos', auth, require('./catalogo.routes'));
-router.use('/auditoria', auth, require('./auditoria.routes'));
-router.use('/reportes', auth, require('./reportes.routes'));
-router.use('/configuracion', auth, require('./configuracion.routes'));
-router.use('/replicacion', auth, require('./replicacion.routes'));
+router.use('/usuarios', auth, franja, require('./usuario.routes'));
+router.use('/accesos', auth, franja, require('./accesos.routes'));
+router.use('/roles', auth, franja, require('./rol.routes'));
+router.use('/catalogos', auth, franja, require('./catalogo.routes'));
+router.use('/auditoria', auth, franja, require('./auditoria.routes'));
+router.use('/reportes', auth, franja, require('./reportes.routes'));
+router.use('/configuracion', auth, franja, require('./configuracion.routes'));
+router.use('/replicacion', auth, franja, require('./replicacion.routes'));
 
 module.exports = router;